Extrat loader runtime communication channel logic to its own crate
