+++ experimental/uefi/channel/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "channel"
+version = "0.1.0"
+authors = ["Juliette Pretot <julsh@google.com>"]
+edition = "2021"
+license = "Apache-2.0"
+
+[features]
+default = ["ring-crypto"]
+ring-crypto = ["oak_remote_attestation_sessions/ring-crypto"]
+
+[dependencies]
+anyhow = { version = "*", default-features = false }
+oak_remote_attestation_sessions = { path = "../../../remote_attestation_sessions", default-features = false }
+ciborium-io = { version = "*", default-features = false }
